// src/app/models/ParkResponse.ts
export interface ParkResponse{
    id: string;
    code: string;
    name: string;
    cnpj: string;
    description: string;
    zipCode: string;
    street: string;
    full: string;
    city: string;
    county: string;
    state: string;
    country: string;
    lat: string;
    lng: string;
    orgShortName: string;
    orgName: string;
    orgCnpj: string;
    orgAddress: string;
    email: string;
    phone: string;
    rolePerson: string;
    relationsManagerId: string;
    hubeesSpots: number;
    businessHours: {
      end: number;
      start: number;
      enabled: boolean;
      dayOfWeek: number;
    };
    partnerParkingId: string;
    showInTheMap: boolean;
    hubeesSubscriptionEnabled: boolean;
    singleStayPaymentEnabled: boolean;
    initialCreation: boolean;
    active: boolean;
    paymentDay: number;
    createdAt: string;
    updatedAt: string;
    discount: {
      id: string;
      parkId: string;
      details: string;
      discountType: string;
      active: boolean;
      amount: string;
      percentage: number;
      createdAt: string;
      updatedAt: string;
    };
    parkingFeature: {
      id: string;
      parkId: string;
      softwareType: string;
      physicalParking: string;
      rainProtection: string;
      createdAt: string;
      updatedAt: string;
    };
}